+++ demo/src/index.js
@@ -1,3 +1,8 @@
-import renderApp from '../../src/app/app';
+import { renderApp } from '../../src/app/app';
 
-renderApp('app');
+const urlParams = new URLSearchParams(window.location.search);
+if (urlParams.has('small')) {
+  renderApp('small-app', 100);
+} else {
+  renderApp('full-app');
+}