+++ CHANGELOG.md
@@ -11,11 +11,11 @@
 - Union, intersection, and complement of cell sets.
 - View all set descendants at a particular level.
 - More public demos.
+- Export `validateAndRender` for use as component.
 ### Changed
 - Gave up on CSS modules. We'll still split out CSS into separate files, but the extra name-munging was just confusing.
 - Crosshairs are now full height and width.
 - S3 data is now organized by dataset.
-- Not making a UMD bundle any longer: It was very large, and slow.
 
 ## [0.0.17](https://www.npmjs.com/package/vitessce/v/0.0.17) - 2019-07-19
 ### Added